+++ better_proto_generator/lib/src/proto_from_cache.dart
@@ -12,6 +12,8 @@ class ProtoFromCache implements Builder {
 
   @override
   Future<void> build(BuildStep buildStep) async {
+    print('BUILD STEP 2: ProtoFromCache');
+
     final inputId = buildStep.inputId;
     final content = await buildStep.readAsString(inputId);
     final copyId = inputId.changeExtension('.proto');